safari-534.53-branch/LayoutTests/fast/canvas/webgl/texture-complete-expected.txt

 
Checks that a texture that is not -texture-complete- does not draw if filtering needs mips

PASS texture that is not -texture-complete- when TEXTURE_MIN_FILTER not NEAREST or LINEAR should draw with 0,0,0,255
PASS successfullyParsed is true

TEST COMPLETE
